Major Update
Updating the example project to be compliant with current version of KiCAD and versions 6.0 and above. This method/protocol was developed using KiCAD 4.0 nightly builds throughout the 2nd half of 2016. The old examples have been moved to a folder called legacy. I will be updating this over the next few weeks as I have free time to spend on it and get an example for every library type I have.